+++ rqueue-core/src/main/java/com/github/sonus21/rqueue/core/RqueueMessageTemplate.java
@@ -58,6 +58,10 @@ public interface RqueueMessageTemplate {
   List<RqueueMessage> getAllMessages(
       String queueName, String processingQueueName, String delayQueueName);
 
+  Long getScore(String zsetName, RqueueMessage message);
+
+  boolean addScore(String zsetName, RqueueMessage message, long delta);
+
   MessageMoveResult moveMessageListToList(
       String srcQueueName, String dstQueueName, int numberOfMessage);