+++ rqueue-core/src/main/java/com/github/sonus21/rqueue/listener/RqueueExecutor.java
@@ -94,6 +94,7 @@ class RqueueExecutor extends MessageContainerBase {
               beanProvider.getRqueueConfig(),
               beanProvider.getRqueueMessageMetadataService(),
               beanProvider.getRqueueJobDao(),
+              beanProvider.getRqueueMessageTemplate(),
               queueDetail,
               messageMetadata,
               rqueueMessage,