+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Pause/Unpause queue
 * Batch message fetching
 * Default queue priority to WEIGHTED
+* Added an API to update the visibility timeout of running job
 
 ## [2.7.0] - 13-Apr-2021